review: useQuery hooks에서 http 통신 다시 service로 분리
